+++ packages/config/eslint/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @stallning/eslint
+# @portfolio/eslint
 
 ## 1.2.0
 
@@ -10,7 +10,7 @@
 
 - cf7f713: chore(config): :wrench: use catalog to share package version
 - cf7f713: chore(config): :wrench: add typescript
-- cf7f713: chore(global): :truck: rename devoid to stallning
+- cf7f713: chore(global): :truck: rename devoid to portfolio
 - cf7f713: fix(config): :bug: skip husky on release action
 - cf7f713: chore(config): :wrench: add eslint oxlint and prettier
 - cf7f713: chore(config): :wrench: add turbo
@@ -23,7 +23,7 @@
 
 ### Patch Changes
 
-- 248922c: chore(global): :truck: rename devoid to stallning
+- 248922c: chore(global): :truck: rename devoid to portfolio
 - 248922c: chore(config): :wrench: add turbo
 - 248922c: chore(config): :wrench: add eslint oxlint and prettier
 - 248922c: chore(config): :wrench: use catalog to share package version