+++ packages/config/typescript/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @stallning/typescript
+# @portfolio/typescript
 
 ## 1.2.0